python/requirements.txt
streamlit>=1.13.0
openai>=1.0.0
altair>=4.2.0
duckduckgo_search>=6.2.0,<7.0.0
playsound>=1.2.2
python-dotenv>=0.19.0
fastapi>=0.115.12
youtube_search>=2.1.2
langchain>=0.3.25
langchain-community>=0.3.24
Pillow>=10.1.0
numpy>=1.26.2

# 🎬 비디오 처리를 위한 추가 라이브러리
moviepy>=1.0.3
imageio>=2.31.1
imageio-ffmpeg>=0.4.9

# 오디오 처리
pydub>=0.25.1

# 이미지 처리

# ✅ HTTP 요청을 위한 라이브러리 추가
requests>=2.31.0
opencv-python>=4.11.0.86
